使用流量大到2G的时候会溢出，所以把d字段改成float的，顺便为了好看，把u字段也改了

+++ app/Models/User.php
@@ -21,8 +21,8 @@ class User extends Model
 
     protected $casts = [
         "t" => 'int',
-        "u" => 'int',
-        "d" => 'int',
+        "u" => 'float',
+        "d" => 'float',
         "port" => 'int',
         "transfer_enable" => 'float',
         "enable" => 'int',
